+++ NEWS.md
@@ -12,6 +12,7 @@
 - add function to get chemical formula
 - fix wrong object name in getmzrt/getmzrt2
 - add options to export EIC object
+- add function to screen organohalogen compounds
 
 # enviGCMS 0.5.0